pyproject.toml
[project]
name = "sick-annual-leave-tracker"
version = "0.1.0"
description = "Tracks Annual and Sick Leave"
authors = [
    {name = "Devon",email = "ds.pydev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "tkcalendar (>=1.6.1,<2.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
